+++ src/template.js
@@ -2,10 +2,12 @@
 
 const visit = require('unist-util-visit');
 const nodeToString = require('hast-util-to-string');
-const refractor = require('refractor');
+<%= refractor %>
 
 module.exports = options => {
   options = options || {};
+
+  <%= languages %>
 
   return tree => {
     visit(tree, 'element', visitor);
@@ -24,9 +26,8 @@ module.exports = options => {
 
     let result;
     try {
-      parent.properties.className = (parent.properties.className || []).concat(
-        'language-' + lang
-      );
+      parent.properties.className = (parent.properties.className || [])
+        .concat('language-' + lang);
       result = refractor.highlight(nodeToString(node), lang);
     } catch (err) {
       if (options.ignoreMissing && /Unknown language/.test(err.message)) {